backend/requirements.txt
# Backend Requirements for Resume Parser with Supabase
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.0
spacy==3.7.2
supabase==2.3.4
python-dotenv==1.0.0

# For spaCy language model, run after installing:
# python -m spacy download en_core_web_sm
# Or for better results:
# python -m spacy download en_core_web_md
